+++ src/ipyelk/tools/tool.py
@@ -1,5 +1,6 @@
 # Copyright (c) 2024 ipyelk contributors.
 # Distributed under the terms of the Modified BSD License.
+from __future__ import annotations
 
 import asyncio
 from typing import Callable
@@ -12,18 +13,22 @@ from ..pipes import Pipe
 
 
 class Tool(W.Widget):
+    """An interactive element to control a diagram."""
+
     tee: Pipe = T.Instance(Pipe, allow_none=True).tag(
         sync=True, **W.widget_serialization
     )
-    on_done = T.Any(allow_none=True)  # callback when done
+    on_done: Callable | None = T.Callable(
+        default_value=None, allow_none=True
+    )  # callback when done
     disable = T.Bool(default_value=False).tag(sync=True, **W.widget_serialization)
     reports = TypedTuple(T.Unicode(), kw={})
     _task: asyncio.Future = None
     ui = T.Instance(W.DOMWidget, allow_none=True)
     priority = T.Int(default_value=10)
-    _on_run_handlers = W.CallbackDispatcher()
+    _on_run_handlers = T.Instance(W.CallbackDispatcher, kw={})
 
-    def handler(self, *args):
+    def handler(self, *_: object) -> asyncio.Future:
         """Handler callback for running the tool"""
         # canel old work if needed
         if self._task:
@@ -38,6 +43,8 @@ class Tool(W.Widget):
         if self.tee:
             self.tee.inlet.flow = self.reports
 
+        return self._task
+
     async def run(self):
         raise NotImplementedError
 
@@ -50,7 +57,7 @@ class Tool(W.Widget):
         Parameters
         ----------
         remove: bool (optional)
-            Set to true to remove the callback from the list of callbacks.
+            set to true to remove the callback from the list of callbacks.
 
         """
         self._on_run_handlers.register_callback(callback, remove=remove)
@@ -73,7 +80,7 @@ class ToolButton(Tool):
     :param handler: Called when button is pressed.
     """
 
-    handler: Callable = T.Any(allow_none=True)
+    handler: Callable | None = T.Callable(default_value=None, allow_none=True)
     description: str = T.Unicode(default_value="")
 
     @T.default("ui")